+++ src/Styleguide/Components/Artist/__stories__/CurrentEvent.story.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
 import { Section } from "../../../Utils/Section"
-import { CurrentEvent, LargeCurrentEvent } from "../Components/CurrentEvent"
+import { CurrentEvent, LargeCurrentEvent } from "../CurrentEvent"
 
 storiesOf("Styleguide/Artist", module).add("Current Event", () => {
   return (
