+++ src/Styleguide/Components/Artwork/OtherAuctions.tsx
@@ -1,8 +1,8 @@
 import React, { ReactNode } from "react"
-import { Grid, Row, Col } from "../Elements/Grid"
-import { Flex } from "../Elements/Flex"
-import { Button } from "../Elements/Button"
-import { Responsive } from "../Utils/Responsive"
+import { Grid, Row, Col } from "../../Elements/Grid"
+import { Flex } from "../../Elements/Flex"
+import { Button } from "../../Elements/Button"
+import { Responsive } from "../../Utils/Responsive"
 import { Serif } from "@artsy/palette"
 import { take } from "lodash"
